+++ common/src/main/java/me/codexadrian/tempad/api/options/TempadOption.java
@@ -2,6 +2,7 @@ package me.codexadrian.tempad.api.options;
 
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -37,4 +38,6 @@ public abstract class TempadOption {
     public boolean depletesDurability() {
         return false;
     }
+
+    public void tick(ItemStack stack, Entity entity) {};
 }