+++ common/src/main/java/me/codexadrian/tempad/api/options/impl/TimerOption.java
@@ -7,6 +7,7 @@ import me.codexadrian.tempad.common.utils.TeleportUtils;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
@@ -18,26 +19,24 @@ import java.time.temporal.ChronoUnit;
 import java.util.List;
 
 public class TimerOption extends TempadOption {
+    public static final String ID = "Timer";
+
     @Override
     public boolean canTimedoorOpen(Player player, ItemStack stack) {
-        if (stack.getTag() != null && stack.getTag().contains(Tempad.TIMER_NBT)) {
-            long cooldownTimeTag = stack.getTag().getLong(Tempad.TIMER_NBT);
-            Instant cooldownTime = Instant.ofEpochSecond(cooldownTimeTag);
-            return Instant.now().isAfter(cooldownTime);
-        }
-        return true;
+        return !stack.getOrCreateTag().contains(ID);
     }
 
     @Override
     public void onTimedoorOpen(Player player) {
-        TeleportUtils.findTempad(player).getOrCreateTag().putLong(Tempad.TIMER_NBT, Instant.now().plusSeconds(TempadOptionApi.getFuelCost(TeleportUtils.findTempad(player))).getEpochSecond());
+        ItemStack stack = TeleportUtils.findTempad(player);
+        stack.getOrCreateTag().putLong(ID, TempadOptionApi.getFuelCost(stack) * 20L);
     }
 
     @Override
     public void addToolTip(ItemStack stack, Level level, List<Component> components, TooltipFlag flag) {
         long cooldown = timeLeft(stack);
         if (cooldown > 0) {
-            components.add(Component.translatable("tooltip.tempad.timeleft", Component.literal(DurationFormatUtils.formatDuration(cooldown, "mm:ss", true)).withStyle(ChatFormatting.AQUA)).withStyle(ChatFormatting.GRAY));
+            components.add(Component.translatable("tooltip.tempad.timeleft", Component.literal(DurationFormatUtils.formatDuration(cooldown * 1000, "mm:ss", true)).withStyle(ChatFormatting.AQUA)).withStyle(ChatFormatting.GRAY));
         } else {
             components.add(Component.translatable("tooltip.tempad.fullycharged").withStyle(ChatFormatting.AQUA));
         }
@@ -51,24 +50,26 @@ public class TimerOption extends TempadOption {
 
     @Override
     public double getPercentage(ItemStack stack) {
-        return (TempadOptionApi.getFuelCost(stack) * 1000.0 - timeLeft(stack)) / (TempadOptionApi.getFuelCost(stack) * 1000.0);
+        return Math.max(0, 1 - (double) timeLeft(stack) / (double) TempadOptionApi.getFuelCost(stack));
     }
 
     public static long timeLeft(ItemStack stack) {
-        if (stack.getTag() != null && stack.getTag().contains(Tempad.TIMER_NBT)) {
-            long cooldownTimeTag = stack.getTag().getLong(Tempad.TIMER_NBT);
-            Instant cooldownTime = Instant.ofEpochSecond(cooldownTimeTag);
-            return Instant.now().until(cooldownTime, ChronoUnit.MILLIS);
+        if (stack.getTag() != null && stack.getTag().contains(ID)) {
+            return stack.getTag().getInt(ID) / 20;
         }
         return 0;
     }
 
-    public static long getTimerNBT(ItemStack stack) {
-        if (stack.getTag() != null) {
-            if (stack.getTag().contains(Tempad.TIMER_NBT)) {
-                return stack.getTag().getLong(Tempad.TIMER_NBT);
+    @Override
+    public void tick(ItemStack stack, Entity entity) {
+        // tick the timer
+        if (stack.getTag() != null && stack.getTag().contains(ID)) {
+            long cooldownTimeTag = stack.getTag().getLong(ID);
+            if (cooldownTimeTag == 0) {
+                stack.getTag().remove(ID);
+            } else {
+                stack.getTag().putLong(ID, cooldownTimeTag - 1);
             }
         }
-        return 0;
     }
 }