+++ common/src/main/java/me/codexadrian/tempad/common/compat/botarium/TempadEnergyContainer.java
@@ -1,0 +1,19 @@
+package me.codexadrian.tempad.common.compat.botarium;
+
+import earth.terrarium.botarium.common.energy.impl.SimpleEnergyContainer;
+
+public class TempadEnergyContainer extends SimpleEnergyContainer {
+    public TempadEnergyContainer(long maxCapacity) {
+        super(maxCapacity);
+    }
+
+    @Override
+    public long maxInsert() {
+        return Integer.MAX_VALUE;
+    }
+
+    @Override
+    public long maxExtract() {
+        return Integer.MAX_VALUE;
+    }
+}